### 2019-04-14 #1376 @whedon
[PRE REVIEW]: labjack-controller: Fast Realtime Data Collection with Labjack T-Series DAQs in Python
**Submitting author:** @nyctanthous (<a href="http://orcid.org/0000-0002-1240-5385">Benjamin Montgomery</a>)
**Repository:** <a href="https://github.com/university-of-southern-maine-physics/labjack-controller" target ="_blank">https://github.com/university-of-southern-maine-physics/labjack-controller</a>
**Version:** 0.3
