### 2019-04-14 #1376 @labarba
@Nyctanthous — have a look at the tabs on the metadata section of your paper. Looks like some weird characters are messing up the paper compilations.
