### 2019-04-14 #1376 @labarba
Also, after a quick look at your paper, I'd say it doesn't quite meet [our expectations](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). In particular, see if you can address these better:
- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- A clear statement of need that illustrates the purpose of the software for research (this should include mention of other software that may address the same need, in a different way, for example)
