### 2019-04-14 caa77ee
Stripped out tab characters in header
