### 2019-04-14 #1376 @Nyctanthous
@labarba Thanks for the heads-up, I'll have the paper revised in the next 24 hours.
