### 2019-04-14 f3ea143
Paper restructure
Clarification and expansion of the paper to better comply with JOSS' guidelines, as noted at the head of openjournals/joss-reviews#1376
