+++ paper.bib
@@ -9,3 +9,9 @@
   howpublished = {\url{https://labjack.com/products}),
   note = {Accessed: 2019-04-08}
 }
+
+@misc{ni,
+  title = {nidaqmx-python},
+  howpublished = {\url{https://github.com/ni/nidaqmx-python/}),
+  note = {Accessed: 2019-04-14}
+}
