### 2019-04-14 0426623
Paper header correction
Addressed formatting issue noted in the build log of openjournals/joss-reviews#1376#issuecomment-483060915
