### 2019-04-14 4d82112
Paper affiliations revision
`Affiliation` element in paper header was incorrect, see openjournals/joss-reviews#1376
