+++ paper.md
@@ -9,10 +9,13 @@ tags:
 authors:
   - name: Benjamin A. Montgomery
     orcid: 0000-0002-1240-5385
-    affiliation: University of Southern Maine
+    affiliation: 1
   - name: Paul A. Nakroshis
     orcid: 0000-0003-1887-354X
-    affiliation: University of Southern Maine
+    affiliation: 1
+affiliations:
+- name: University of Southern Maine
+  index: 1
 date: 5 April 2019
 bibliography: paper.bib
 ---
