### 2019-04-14 1dc27e7
Bibliography typo correction
Again, see build logs in openjournals/joss-reviews#1376
