+++ paper.bib
@@ -6,12 +6,12 @@
 
 @misc{Labjack,
   title = {Labjack},
-  howpublished = {\url{https://labjack.com/products}),
+  howpublished = {\url{https://labjack.com/products}},
   note = {Accessed: 2019-04-08}
 }
 
 @misc{ni,
   title = {nidaqmx-python},
-  howpublished = {\url{https://github.com/ni/nidaqmx-python/}),
+  howpublished = {\url{https://github.com/ni/nidaqmx-python/}},
   note = {Accessed: 2019-04-14}
 }
