### 2019-04-15 c225fa0
Added explanation of applicability
Edited to make it clear why this is a very valuable package for researchers interested in getting something done quickly, reliably, and inexpensively.
