### 2019-04-16 #1376 @usethedata
@Nyctanthous In your document, I see "n.d." Is that for "no DOI"?  Sorry to be slow in responding -- some challenges in the day job.
