### 2019-04-16 #1376 @Nyctanthous
@usethedata No problem. I expect n.d. indicates "no date" recorded for authorship; I'm using the `misc` entry in the bibliography because, at this time, all references are websites.

I am not aware of any papers on the libraries that are similar/related to this one. ([minimal-modbus][0], [nidaqmx-python][1], [openDAQ][2], [Lawson Lab's 301/302 library][3], as some popular examples). If are any, however, it would be great to add them.

[0]: https://minimalmodbus.readthedocs.io/en/master/
[1]: https://nidaqmx-python.readthedocs.io/en/latest/
[2]: https://github.com/openDAQ/python-opendaq
[3]: http://www.lawsonlabs.com/specExternal301_Developer.html
