### 2019-05-06 4d412ec
Added NumPy referencing for docs
