### 2019-05-12 #1376 @Nyctanthous
It looks like @ixjlyons would be a good fit, given their preference for data acquisition and assorted electronic systems experience.
@lucask07 also looks like they should have a significant experience in this field, as this library is all about interacting with instruments, usually sensors, and experimental data acquisition.
