### 2019-05-12 #1376 @labarba
@ixjlyons, @lucask07 👋 — would either (or both) of you be willing to contribute a review for this JOSS submission? Title: "labjack-controller: Fast Realtime Data Collection with Labjack T-Series DAQs in Python" ... Let us know!
