### 2019-05-13 #1376 @lucask07
Hi @labarba
Sure, I have a labjack and so this is definitely a good fit.
