### 2019-05-13 #1376 @ixjlyons
I don't have a LabJack device, but I have experience with data acquisition so I could do some code review and the usual checklist items (and I'm willing to do the review if that's sufficient). Alternatively, I know @moorepants is on the reviewer list and has LabJack hardware, though I don't know if he has one of the T series devices or bandwidth to do a review.
