### 2019-05-28 #1448 @ixjlyons
My review:

labjack-controller is a Python library which aims to hide some of the low-level details of collecting data from a LabJack device. The interface is well thought-out and the documentation contains enough information to get started and make full use of the library. The code is very clean and seems to be written well with good error handling (and nice use of type hints!). The paper presents both the motivations for the library and the research purpose clearly.

There were a minor issues I noticed in the documentation:

- Overview page:
    - Requirements section:
        - In the note on LJM, typo: "The LJM library is provided a a..."
    - Installation section:
        - The installation steps miss `pip install .`. It's mentioned below but I think the note below should say you can omit `pip install -r requirements.txt` if you already have LJM installed.
        - There is no `requirements.txt` (use the one in `docs/`?)
- Demo notebook 1
    - Step 1 section
        - Typo "...there will be a notably offset..."
        - Typo "Additionally, if you are perofrming..."
- API Documentation
    - There's an empty "Module contents" section at the bottom
    - `LabJackReader.collect_data` example says 50kHz sample rate but shows 10 kHz.
- Contributing guidelines are missing (part of community guidelines).

And on to a couple bigger issues. After reading the docs through, it's still not perfectly clear to me the interaction between `scan_rate` and `scans_per_read`. From the descriptions given, `scan_rate` sounds like sample rate (i.e. how rapidly the LabJack actually samples the analog inputs or effectively does so [i.e. if it supersamples internally, not sure]), and `scans_per_read` sounds like it controls the number of samples returned on each update from the device. If this is the case, consider using "update rate," "read rate," or maybe "frame rate" instead of "sample rate" in the `scans_per_read` docstring -- "sample rate" to me usually indicates the rate of sampling the analog signal.

The use of the term "realtime" to me implies the library is set up to continuously stream data, such as reading a frame of samples at a time and continuously updating a live plot with the collected data, say, every second. Looking at the documentation and code, it seems the library is more for doing what I would maybe call "data logging." The only `LabjackReader` method for recording data is `collect_data`, which opens and closes the stream on each call, so it doesn't look to me like it can be used for continuous streaming where the user needs to periodically work with the data being streamed. This could be implemented with a callback mechanism where the user can pass a callback function to `collect_data` which receives each frame as it comes in, for example. Currently, the user could work with `LJMLibrary` to achieve this, but this is fairly low-level functionality which I think labjack-controller aims to hide. I don't think it's *necessary* to support this kind of functionality for the library to be useful, but after reading the paper, I'd expect that capability to be implemented. I would recommend that you either implement an interface for continuous streaming or make it clear in the paper/docs that the main functionality is collecting data for a specific amount of time.

Overall, labjack-controller is a well-written library and I think it has sufficient general scientific research purpose to fit in to JOSS. I recommend accepting so long as the notes above are taken into consideration.
