### 2019-05-28 5d46122
Create CONTRIBUTING.md
