### 2019-05-28 72209f7
Typo corrected, installation instructions updated
