### 2019-05-28 #1448 @Nyctanthous
@ixjlyons Thank you, you caught some things that totally slipped by me. All of the minor issues you mentioned should be fixed in the next 24 hours or so.

Adding a callback as a kwarg is trivial; I'll look into a solution to make the realtime bit more accessible. Currently, realtime functionality requires parallel processing; the Labjack vomits data all over the place, as soon as you start streaming and if you spend too much time in between acquiring frames, it will overflow all internal buffers and the stream will die. I expect I can hide this issue from the user by maintaining a thread pool and spinning off new callbacks from the pool.

I'll also clear up the `scans_per_read` / `scan_rate` confusion in the documentation and code. `scan_rate` is sampling frequency, `scans_per_read` represents the limitations of inter-device communication; when you send information from the labjack to your computer, it does it in batches, where each "scan" is a complete row of data. `scans_per_read` determines this batch size.

I expect to have everything fixed up by this weekend.
