+++ docs/labjackcontroller.rst
@@ -9,11 +9,3 @@ labjackcontroller.labtools module
     :undoc-members:
     :show-inheritance:
 
-
-Module contents
----------------
-
-.. automodule:: labjackcontroller
-    :members:
-    :undoc-members:
-    :show-inheritance:
