### 2019-05-29 #1448 @lucask07
This is a very nicely written software package and certainly fills a need. I'm a LabJack user and realize that the cost and ease of use (as compared to say, an NiDAQ) is great. Keep up the nice work!

### Review checkboxes that are yet to be ticked:

- **Community Guidelines** Missing a note on how to ask for support
- **References** in the paper are all cited as "n.d." This should be fixed somehow.
- **Functionality: Have the functional claims of the software been confirmed?** need clarification or rewording as it relates to 'realtime'. Or a demonstration (with a function generator) that realtime is achieved.

### Other Notes:

Unfortunately, I have a U6 LabJack (not a T series) so am not able to test this package with actual hardware. Note that for the U series LabJack has published a Python API: [U series Python API](https://github.com/labjack/LabJackPython).

#### The title of the paper does not seem to accurately reflect what the software provides.

- By "Fast" do you mean:

1. that the user can get the software up and running quickly?
2. or that the data transfer rate to the host computer is maximized in some way?

- "Realtime" is a very specific term and often implies that data is sampled at clock rates entirely set by hardware and no (non-realtime) operating system is involved. I suspect with this software the data acquired is in effect "realtime" assuming that only **one** LabJack to host data transfer occurs. If, however, multiple transfers occur there will be indeterminate gaps in time set by the latency of the LabJack to host transfer. A way around this so that this package continuously streams data at an exact sampling rate would be a great addition. Alternatively, a demonstration with a sine-wave function generator input (at a frequency of ~200 Hz) may help explain to the user:
1. the number of sampling points that can be retrieved with no gaps and
2. the number of sampling points that produce a gap in the time sequence and the size of that gap.

#### A few other notes, ideas, and questions

- The paper references "real-time reaction to sensor readings". I don't follow how this is happening. Again, to me, "real-time" implies no interaction with a non-realtime operating system. Please explain or remove from the paper.

- I'm noting some confusion with the setting mode of analog streaming versus the settling time. I see both of these variables and, at times, one may be mixed up for the other:
``stream_setting`` versus ``stream_settling``

For example, within the docstring for ``_setup``
```
        stream_settling: int, optional
            See official LabJack documentation.
```
I'd expect that a settling time would be a float. Could you please check this? And add to the documentation the streaming mode(s) that you intend to support?

- Is there a plan to implement an interface to the digital I/O?
