### 2019-05-29 Nyctanthous/labjack-controller#4 @nyctanthous
JOSS Reviewer Revisions A
As per https://github.com/openjournals/joss-reviews/issues/1448#issuecomment-496911911,

## Wording
- [x] Clarify usage of "Fast" in all documents.
- [x] Remove / clarify usage of the term "Real-Time" in all documents.
- [x] Fix references in paper.

## Refactoring
- [x] Refactor `stream_settling` into the `LabjackReader` settings function.
- [x] Remove `stream_setting`.

## Demos
- [x] Clarify support for digital I/O, give examples.
