### 2019-05-29 Nyctanthous/labjack-controller#5 @nyctanthous
JOSS Reviewer Revisions B
As per https://github.com/openjournals/joss-reviews/issues/1448#issuecomment-496365163,

## Naming
- [x] Rename `scan_rate` and `scans_per_read`.
- [x] Remove usage of the term "real-time"
- [x] Add a callback as a kwarg to `collect_data` to make interacting with live data easier.
