### 2019-05-30 9165177
Document rephrasing
Removal of phrases "realtime" and "fast". Initial implementation of thread pool for callback. Addresses some issues presented in #4, #5.
