+++ labjackcontroller/labtools.py
@@ -3,6 +3,8 @@ from labjack.ljm import constants as ljm_constants, \
                         errorcodes as ljm_errorcodes
 from labjack.ljm.ljm import LJMError
 
+from multiprocessing import Process, Pool
+
 import numpy as np
 import pandas as pd
 from typing import List, Tuple, Union
@@ -928,7 +930,7 @@ class LabjackReader(object):
                       " stream running.")
             pass
 
-    def _setup(self, inputs, inputs_max_voltages, stream_settling, resolution,
+    def _setup(self, inputs, inputs_max_voltages, resolution,
                scan_rate, scans_per_read=-1) -> Tuple[int, int]:
         """
         Set up a connection to the LabJack for streaming
@@ -941,8 +943,6 @@ class LabjackReader(object):
         inputs_max_voltages: sequence of real values
             Maximum voltages corresponding element-wise to the channels
             listed in inputs.
-        stream_settling: int, optional
-            See official LabJack documentation.
         resolution: int, optional
             See official LabJack documentation.
         scan_rate: int
@@ -979,12 +979,12 @@ class LabjackReader(object):
 
         # Next, verify the scans/read.
         if scans_per_read == -1:
-            scans_per_read = scan_rate
+            scans_per_read = int(scan_rate / 2)
         elif scans_per_read > scan_rate:
             warnings.warn("Maximum valid scan/read rate is larger than"
                           " the scan rate. Setting to be equal to the scan"
                           " rate.", UserWarning)
-            scans_per_read = scan_rate
+            scans_per_read = int(scan_rate / 2)
 
         # If a packet is lost, don't try and get it again.
         self._ljm_reference.modify_settings(retry_on_transaction_err=False)
@@ -1010,9 +1010,6 @@ class LabjackReader(object):
 
         names.extend([element + "_RANGE" for element in ain_inputs])
         values.extend(inputs_max_voltages)
-
-        self.modify_settings(stream_settling_time=stream_settling,
-                             stream_resolution=resolution)
 
         # Write the analog inputs' negative channels (when applicable),
         # ranges, stream settling time and stream resolution configuration.
@@ -1242,7 +1239,6 @@ class LabjackReader(object):
     def find_max_freq(self,
                       inputs: List[str],
                       inputs_max_voltages: List[float],
-                      stream_setting=0,
                       resolution=0,
                       verbose=True,
                       max_buffer_size=0,
@@ -1259,8 +1255,6 @@ class LabjackReader(object):
         inputs_max_voltages : sequence of real values
             Maximum voltages corresponding element-wise to the channels
             listed in inputs.
-        stream_setting : int, optional
-            See official LabJack documentation.
         resolution : int, optional
             See official LabJack documentation.
         verbose : str, optional
@@ -1328,7 +1322,6 @@ class LabjackReader(object):
                     self.open(verbose=False)
                     scan_rate, sample_rate = self._setup(inputs,
                                                          inputs_max_voltages,
-                                                         stream_setting,
                                                          resolution,
                                                          med_rate,
                                                          scans_per_read=scans_per_read)
@@ -1467,18 +1460,16 @@ class LabjackReader(object):
                            (Fore.RED if ljm_buffer_size > MAX_LJM_BUFFERSIZE else Fore.RESET) + str(ljm_buffer_size) + Fore.RESET,
                            (Fore.RED if num_skips > 0 else Fore.RESET) + str(num_skips) + Fore.RESET))
 
-    # TODO: stream_setting needs to be renamed and abstracted away, see
-    # STREAM_SETTLING_US
-    # at https://labjack.com/support/datasheets/t-series/communication/stream-mode
     def collect_data(self,
                      inputs: List[str],
[…]
