### 2019-06-03 b290bf1
Renamed scan variables, finished callback function
Closes #5
