+++ tests/labjackcontroller_test.py
@@ -11,8 +11,8 @@ def get_ljm_devices():
 
 @pytest.fixture(scope='session')
 def ljm_all_channels():
-    channels = [*["AIN" + str(num) for num in range(0, 3)],
-                *["DIO" + str(num) for num in range(0, 3)]]
+    channels = [*["AIN" + str(num) for num in range(0, 1)],
+                *["DIO" + str(num) for num in range(0, 1)]]
     return [subset for num_ch in range(1, len(channels) + 1)
             for subset in itertools.combinations(channels, num_ch)]
 
@@ -72,13 +72,14 @@ def test_open_close_status(get_ljm_devices):
 def test_collect_data_gathering(get_ljm_devices, ljm_all_channels, resolution,
                                 frequency):
     # Duration of each test, in seconds.
-    duration = 10
+    duration = 1
 
     for device_args in get_ljm_devices:
         curr_device = LabjackReader(*device_args[:3])
 
         for channel_config in ljm_all_channels:
             # Iterate through some amount of channels.
+            print("Starting with channels", channel_config)
             tot_time, num_skips = curr_device \
                 .collect_data(channel_config,
                               [10.0] * len(channel_config),
@@ -93,12 +94,12 @@ def test_collect_data_gathering(get_ljm_devices, ljm_all_channels, resolution,
             assert tot_time > duration - 0.01
 
 
-@pytest.mark.parametrize("inputs, inputs_max_voltages, seconds, scan_rate,"
+@pytest.mark.parametrize("inputs, inputs_max_voltages, seconds, frequency,"
                          " scans_per_read, resolution,"
                          " expected_err",
                          [(["AIN0"], [10.0], 1, 1000, 500, 1, None)])
 def test_collect_data_parameters(get_ljm_devices, inputs, inputs_max_voltages,
-                                 seconds, scan_rate, scans_per_read,
+                                 seconds, frequency, scans_per_read,
                                  resolution, expected_err):
     if not len(get_ljm_devices):
         return
@@ -107,47 +108,68 @@ def test_collect_data_parameters(get_ljm_devices, inputs, inputs_max_voltages,
             with pytest.raises(expected_err):
                 with LabjackReader(*get_ljm_devices[0][:3]) as curr_device:
                     curr_device.collect_data(inputs, inputs_max_voltages,
-                                             seconds, scan_rate,
+                                             seconds, frequency,
                                              scans_per_read=scans_per_read,
                                              resolution=resolution)
     else:
         with LabjackReader(*get_ljm_devices[0][:3]) as curr_device:
             curr_device.collect_data(inputs, inputs_max_voltages,
-                                     seconds, scan_rate,
+                                     seconds, frequency,
                                      scans_per_read=scans_per_read,
                                      resolution=resolution)
 
 
-def test_to_list(get_ljm_devices):
+def test_callbacks(get_ljm_devices):
     for device_args in get_ljm_devices:
         # First test with no data stored.
         curr_device = LabjackReader(*device_args[:3])
 
-        assert curr_device.to_list(mode="all") is None
-        assert curr_device.to_list(mode='relative', num_rows=50) is None
-        assert curr_device.to_list(mode='range', start=17, end=65) is None
+        channels_to_scan = ["AIN0", "AIN2", "AIN4"]
+
+        # Scan for 1 second at 10 Hz.
+        curr_device.collect_data(channels_to_scan, 3 * [10.0], 1, 10,
+                                 callback_function=new_callback)
+
+        # Should return back our 3 data column and 2 time columns.
+        assert np.shape(curr_device.to_array(mode="all")) == (10, 5)
+
+
[…]
