+++ labjackcontroller/labtools.py
@@ -344,7 +344,7 @@ class LJMLibrary(metaclass=Singleton):
         port = ctypes.c_int32(0)
         max_packet_size = ctypes.c_int32(0)
 
-        error = self._ljm_reference.staticlib\
+        error = self.staticlib\
             .LJM_GetHandleInfo(handle,
                                ctypes.byref(device_type),
                                ctypes.byref(connection_type),
@@ -564,7 +564,7 @@ class LJMLibrary(metaclass=Singleton):
         return packet_data, dev_buffer_backlog.value, \
             ljm_buffer_backlog.value
 
-    def stream_start(self, handle: int, scan_list: List[str], scan_rate: float,
+    def stream_start(self, handle: int, scan_list: List[str], frequency: float,
                      scans_per_read: int) -> float:
         """
         Based on the LJM function LJM_eStreamStart. Creates a buffer that the
@@ -578,7 +578,7 @@ class LJMLibrary(metaclass=Singleton):
             connection.
         scan_list : List[str]
             List of addresses ("AIN0", etc) to scan from.
-        scan_rate : float
+        frequency : float
             The rate in Hz that all specified addresses will be read at.
         scans_per_read : int
             When reading data from the stream, this value determines how
@@ -587,7 +587,7 @@ class LJMLibrary(metaclass=Singleton):
 
         Returns
         -------
-        actual_scan_rate: float
+        actual_frequency: float
             The actual frequency the device will scan at.
 
         Raises
@@ -600,7 +600,7 @@ class LJMLibrary(metaclass=Singleton):
 
         self._validate_handle(handle)
 
-        scan_rate = ctypes.c_double(scan_rate)
+        frequency = ctypes.c_double(frequency)
         num_addrs = len(scan_list)
         scan_list = self._names_to_modbus_addresses(scan_list)
         self._ljm_buffer[handle] = scans_per_read * num_addrs
@@ -609,12 +609,12 @@ class LJMLibrary(metaclass=Singleton):
                                                 c_int32(scans_per_read),
                                                 c_int32(num_addrs),
                                                 ctypes.byref(scan_list),
-                                                ctypes.byref(scan_rate))
+                                                ctypes.byref(frequency))
 
         if error != ljm_errorcodes.NOERROR:
             raise LJMError(error)
 
-        return scan_rate.value
+        return frequency.value
 
     def stream_stop(self, handle: int) -> None:
         """
@@ -834,8 +834,8 @@ class LabjackReader(object):
 
         return "LabjackReader('Type': %s, Connection': %s, 'Serial': %i," \
             " 'IP': %s, 'Port': %i)" \
-            % (self._meta_device, self._meta_connection, self._meta_serial_number,
-               self._meta_ip_addr, self._meta_port)
+            % (self._meta_device, self._meta_connection,
+               self._meta_serial_number, self._meta_ip_addr, self._meta_port)
 
     @property
     def connection_status(self):
@@ -931,7 +931,7 @@ class LabjackReader(object):
             pass
 
     def _setup(self, inputs, inputs_max_voltages, resolution,
-               scan_rate, scans_per_read=-1) -> Tuple[int, int]:
+               frequency, scans_per_read=-1) -> Tuple[int, int]:
         """
         Set up a connection to the LabJack for streaming
 
@@ -945,7 +945,7 @@ class LabjackReader(object):
             listed in inputs.
         resolution: int, optional
             See official LabJack documentation.
-        scan_rate: int
+        frequency: int
             Number of times per second (Hz) the device will get a datapoint for
             each of the channels specified.
         scans_per_read: int, optional
@@ -954,37 +954,37 @@ class LabjackReader(object):
 
         Returns
         -------
-        scan_rate : int
[…]
