### 2019-06-03 8d099d7
Better demo of DIO streaming
Addresses parts of #4.
