+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -10,9 +10,8 @@ A clear and concise description of what the bug is.
 **To Reproduce**
 Steps to reproduce the behavior:
 1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
+2. Run command '....'
+3. See error
 
 **Expected behavior**
 A clear and concise description of what you expected to happen.
@@ -21,15 +20,8 @@ A clear and concise description of what you expected to happen.
 If applicable, add screenshots to help explain your problem.
 
 **Desktop (please complete the following information):**
- - OS: [e.g. iOS]
- - Browser [e.g. chrome, safari]
- - Version [e.g. 22]
-
-**Smartphone (please complete the following information):**
- - Device: [e.g. iPhone6]
- - OS: [e.g. iOS8.1]
- - Browser [e.g. stock browser, safari]
- - Version [e.g. 22]
+ - OS: [e.g. Arch Linux]
+ - Python version [e.g. 22]
 
 **Additional context**
 Add any other context about the problem here.
