### 2019-06-03 096b76f
Update bug_report.md
