+++ paper.bib
@@ -1,17 +1,23 @@
 @misc{Lawson,
   title = {Model 301/302 Developer Application Software},
   howpublished = {\url{http://www.lawsonlabs.com/specExternal301_Developer.html#30xLinux}},
+  year = {2017},
+  month = {October},
   note = {Accessed: 2019-04-08}
 }
 
 @misc{Labjack,
   title = {Labjack},
-  howpublished = {\url{https://labjack.com/products}},
-  note = {Accessed: 2019-04-08}
+  howpublished = {\url{https://labjack.com/support/software/installers/ljm}},
+  year = {2019},
+  month = {May},
+  note = {Accessed: 2019-06-03}
 }
 
 @misc{ni,
   title = {nidaqmx-python},
   howpublished = {\url{https://github.com/ni/nidaqmx-python/}},
+  year = {2017}
+  month = {December}
   note = {Accessed: 2019-04-14}
 }
