+++ paper.bib
@@ -17,7 +17,7 @@
 @misc{ni,
   title = {nidaqmx-python},
   howpublished = {\url{https://github.com/ni/nidaqmx-python/}},
-  year = {2017}
-  month = {December}
+  year = {2017},
+  month = {December},
   note = {Accessed: 2019-04-14}
 }
