### 2019-06-03 #1448 @Nyctanthous
@lucask07 Thanks for your review. I believe that, at this point, I have addressed everything you pointed our as incorrect. A few comments explaining what my thinking was at the time:

1. Usage of the word "Realtime". The T4, and T7 are (probably) realtime in the sense that you indicate. However, this package makes no effort to propagate this attribute through the process of TCP or UDP data streaming. At some future point, I agree with you that it would be very neat to add, but I expect that will take a while to develop and perfect. When I used "realtime", I meant it to act like the word "live"; e.g. "live data processing". That has since been corrected, and "realtime" has been completely removed.
2. Usage of the word "Fast". Anecdotally, the package is faster than a one-to-one Python wrapper of the C functions due to a reduction of overhead; the package is set up in a way that we minimize type conversion, and rely heavily on primitive (`ctypes`, `numpy`) types in the first place. With that said, I have removed these claims because they are difficult to quantify - any benefit gets overshadowed by the cost of opening/closing/communicating with the LabJack.
3. `stream_setting` vs `stream_settling`: The "Setting" option does not exist (surprise, surprise), and was solely exists as a typo. Any actual stream setting out there (there aren't that many) you can set with a Labjack object's `modify_setting` function. This has also been cleared up.
4. Interfacing the digital I/O. We already do! You can stream any 16 bit channel; the process is almost exactly the same. I've added examples to the demo files and the documentation.
5. "n.d." references. This stands for "No Date", due to these references being websites with unclear authorship dates. To come up with a date, I looked at the date that the Python bindings for each of these references had been updated on.

@ixjlyons I think I have also addressed all of your points in the manner I outlined previously.

At this point, I expect we can continue the review and editing process.
