### 2019-06-05 #1448 @lucask07
@Nyctanthous looks great, thanks for the thorough response to my critique. Note that I cannot verify the functionality since I don't have the hardware. I now accept this submission. Nice work!
