### 2019-06-09 a390558
Update CONTRIBUTING.md
