+++ .github/ISSUE_TEMPLATE/question-template.md
@@ -1,0 +1,10 @@
+---
+name: Question template
+about: Use when nothing is broken, but there is something that should be clarified
+title: "[QUESTION]"
+labels: question
+assignees: ''
+
+---
+
+
