### 2019-06-09 #1448 @usethedata
@lucask07 and @ixjlyons There are a couple of things still showing open on the review checklists.  If all your issues are taken care of, can you close out the checklists?  I realize that @lucask07 indicates not being able to complete a functional check. @ixjlyons -- were you able to do a functional check?  If not, I'll see if I can find someone around my lab who has the hardware.
