### 2019-06-10 a56e2f3
Switched to PBR for installation, versioning
