+++ setup.py
@@ -1,27 +1,3 @@
-from setuptools import setup
+import setuptools
 
-setup(name='labjackcontroller',
-      description='A helper library to control LabJack devices',
-      version='0.3',
-      python_requires='>=3.5.1',
-      url='https://github.com/Nyctanthous/labjack-controller',
-      author='Ben Montgomery',
-      packages=['labjackcontroller'],
-      license='MIT',
-      classifiers=[
-          'Development Status :: 4 - Beta',
-          'Intended Audience :: Science/Research',
-          'License :: OSI Approved :: MIT License',
-          'Programming Language :: Python :: 3'
-      ],
-      dependency_links=[
-                        'https://github.com/labjack/labjack-ljm-python/tarball/master',
-                       ],
-      install_requires=[
-                        'typing',
-                        'numpy',
-                        'pandas',
-                        'colorama',
-                        'labjack-ljm'
-                       ],
-      zip_safe=False)
+setuptools.setup(setup_requires=['pbr'], pbr=True)
