+++ setup.cfg
@@ -1,0 +1,19 @@
+[metadata]
+name = labjackcontroller
+author = Ben Montgomery
+summary = A helper library to control LabJack T-series devices
+license = MIT
+description-file =
+    README.md
+home-page = http://pypi.python.org/pypi/foobar
+requires-python = >=3.5.1
+classifier = 
+    Development Status :: 4 - Beta
+    Intended Audience :: Science/Research
+    icense :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3
+
+[files]
+packages =
+    labjackcontroller
