### 2019-06-10 13bf41f
Update setup.cfg
