+++ setup.cfg
@@ -10,7 +10,7 @@ requires-python = >=3.5.1
 classifier = 
     Development Status :: 4 - Beta
     Intended Audience :: Science/Research
-    icense :: OSI Approved :: MIT License
+    License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Programming Language :: Python :: 3
 
