### 2019-06-10 1adceef
Update setup.cfg
