+++ setup.cfg
@@ -5,6 +5,7 @@ summary = A helper library to control LabJack T-series devices
 license = MIT
 description-file =
     README.md
+description-content-type = text/markdown
 home-page = http://pypi.python.org/pypi/foobar
 requires-python = >=3.5.1
 classifier = 
