+++ setup.cfg
@@ -11,6 +11,7 @@ requires-python = >=3.5.1
 classifier = 
     Development Status :: 4 - Beta
     Intended Audience :: Science/Research
+    Topic :: Scientific/Engineering
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Programming Language :: Python :: 3
