### 2019-06-10 #1448 @ixjlyons
The changes look good to me and I checked off my remaining boxes. I don't have a LabJack device so I was not able to test the library myself.
