### 2019-06-16 #1448 @usethedata
@Nyctanthous This paper looks ready to go.  Can you please generate an archive version?  Post the archive DOI back in here and let me know the version of the software you archived.  I'll double check things and update the archive and version here.  Thanks.
