### 2019-06-16 #1448 @usethedata
I confirm that the DOI which whedon suggests above is not relevant to this paper.  It is a reference to a physical device, used to raise and lower laboratory equipment, and has nothing to do with the labjack software.
