### 2019-06-16 #1448 @Nyctanthous
I created the release [0.4.1](https://github.com/university-of-southern-maine-physics/labjack-controller/tree/0.4.1) to make Zenodo happy. DOI is 10.5281/zenodo.3247140
