### 2019-06-18 #1448 @danielskatz
👋 @lucask07 - I note one checkbox that you didn't check - can you please do that before I close this review/issue?

(If there's a problem, we can unpublish the paper - sorry I didn't see this until the last minute)
