### 2019-06-18 #1448 @lucask07
Hi @danielskatz
I am not able to check the functionality box because I do not have the hardware (T-series Labjack). I reviewed the code and confirmed the installation procedure. I'd suggest the paper remains published even without that checkbox.
